Update options when new props are received
Updates the `options` when new props are received.

This allows the props to be updated by the user
to change the wave height, amplitude, speed or number of points.

Note: Changing any of the options props will not smoothly animate the change,
but will instead jump straight to the new state.